+++ prince-archiver/prince_archiver/data_archive/handlers.py
@@ -3,12 +3,12 @@ from typing import Callable
 
 from aio_pika.abc import AbstractIncomingMessage
 
-from prince_archiver.db import AbstractUnitOfWork
 from prince_archiver.models import (
     DataArchiveEntry,
     Timestep,
 )
 from prince_archiver.service_layer.messagebus import MessageBus
+from prince_archiver.service_layer.uow import AbstractUnitOfWork
 
 from .dto import UpdateArchiveEntries
