+++ prince-archiver/prince_archiver/watcher/handlers.py
@@ -3,10 +3,10 @@ from datetime import UTC
 
 from arq import ArqRedis
 
-from prince_archiver.db import AbstractUnitOfWork
 from prince_archiver.dto import TimestepDTO
 from prince_archiver.models import Timestep
 from prince_archiver.service_layer.messagebus import AbstractHandler
+from prince_archiver.service_layer.uow import AbstractUnitOfWork
 
 LOGGER = logging.getLogger(__name__)
