+++ prince-archiver/tests/unit/data_archive/test_handlers.py
@@ -5,6 +5,7 @@ from uuid import uuid4
 import pytest
 from aio_pika.abc import AbstractIncomingMessage
 
+from prince_archiver.adapters.repository import AbstractTimestepRepo
 from prince_archiver.data_archive.dto import (
     Archive,
     UpdateArchiveEntries,
@@ -13,9 +14,9 @@ from prince_archiver.data_archive.handlers import (
     SubscriberMessageHandler,
     add_data_archive_entries,
 )
-from prince_archiver.db import AbstractTimestepRepo, AbstractUnitOfWork
 from prince_archiver.models import Timestep
 from prince_archiver.service_layer.messagebus import MessageBus
+from prince_archiver.service_layer.uow import AbstractUnitOfWork
 
 
 @pytest.fixture(name="update_archive_entries_msg")
